+++ NEWS.md
@@ -11,6 +11,7 @@
 ## Changes to functions
 
 * `ggpredict()` now supports prediction intervals for models from *MCMCglmm*.
+* `ggpredict()` gets a `back.transform`-argument, to tranform predicted values from log-transformed responses back to their original scale (the default behaviour), or to allow predictions to remain on log-scale (new).
 
 # ggeffects 0.10.0